feat: 引入GORM重构数据库层
- 添加GORM依赖并替换原生SQL实现
- 创建internal/db/gorm.go提供GORM初始化和自动迁移功能
- 重写models.go添加GORM标签，时间字段改为time.Time类型
- 修改app.go移除*sql.DB依赖，使用全局GORM实例
- 更新service层使用GORM API进行CRUD操作
- 修复handler层返回类型和时间格式转换问题
- 更新前端model.ts处理time.Time类型转换
- 移除旧的sqlite.go文件和相关原生SQL实现

+++ internal/handler/skill.go
@@ -96,7 +96,7 @@ func (h *SkillHandler) GetRecentSkills(limit int) ([]db.Skill, error) {
 }
 
 // CountSkills 统计 Skill 总数
-func (h *SkillHandler) CountSkills() (int, error) {
+func (h *SkillHandler) CountSkills() (int64, error) {
 	return h.skillSvc.CountSkills()
 }
